parser: byte-oriented lexer + AST size diet for ~2x parse speedup
Parse-only on all-packages.nix drops from ~17 ms to ~13 ms wall (~6 ms
hot-loop), with proportionally larger gains on string-heavy files
(hackage-packages.nix: 1.7x). End-to-end format is ~1.15x faster.

Profiling showed three cost centres dominating the parser:

  * Per-char `Vec<char>` indexing and `advance()` in the lexer.
  * `memmove` of large `Ann<Token>` / `Expression` / `Binder` values
    threaded by-value through the recursive descent.
  * Allocator traffic from per-identifier `String`s and the always-empty
    `Vec<Trivium>` carried by every annotation.

This commit attacks all three:

Lexer
  * Drop the up-front `Vec<char>` and the parallel char-index cursor;
    scan `source` byte-wise, decoding UTF-8 only when a multi-byte lead
    byte is observed. `at()`/`skip_hspace()`/`parse_newlines()` become
    plain byte loops.
  * Add `memchr`-based bulk scanners (`scan_until3`, `seek_to`,
    `advance_bytes_no_newline`) and use them for simple/indented string
    bodies and line/block comments, replacing char-by-char copying.
  * Add `fast_ws_trivia` so the common "only whitespace between tokens"
    case bypasses the scratch-vector + `convert_trivia` machinery.
  * Keyword classification dispatches on (len, first byte) so the usual
    "not a keyword" identifier path is a single comparison.

Types
  * `Span` fields are `u32` (16 B instead of 32 B).
  * `Trivia` stores `Option<Box<Vec<Trivium>>>` (8 B; empty = null).
  * `TrailingComment` holds `Box<str>`.
  * `Token` text is `compact_str::CompactString`, eliminating the
    per-identifier heap allocation that dominated AST drop time.
  * `error::Result<T>` boxes `ParseError` so the error variant is
    pointer-sized instead of ~100 B.

Together these shrink `Ann<Token>` from 112 B to 72 B and roughly halve
the large enum variants, which is what cuts the `memmove` share.

Parser
  * `take_and_advance` writes the next lexeme directly into `current`
    instead of a dummy followed by the lexeme; remaining
    `take_current(); advance()` pairs are folded into it.

Also adds `examples/parse_loop.rs` for in-process profiling.

+++ src/error/format.rs
@@ -103,8 +103,8 @@ impl<'a> ErrorFormatter<'a> {
 
             // Show pointer only on the error line
             if line_idx == error_line_idx {
-                let error_col = error.span.start.saturating_sub(line_start_offset);
-                let error_len = (error.span.end - error.span.start).max(1);
+                let error_col = (error.span.start as usize).saturating_sub(line_start_offset);
+                let error_len = (error.span.end - error.span.start).max(1) as usize;
 
                 // Calculate visual column (counting chars not bytes)
                 let visual_col = line_text[..error_col.min(line_text.len())].chars().count();
